account/requirements.txt
django
djangorestframework
django-cors-headers
django-otp
psycopg2
Pillow
pyotp
pyjwt